my-weather-project/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
psycopg2-binary
requests
streamlit
pandas
matplotlib
plotly
seaborn
altair